+++ framy_generator/lib/framy_generator.dart
@@ -6,8 +6,10 @@ import 'package:framy_annotation/framy_annotation.dart';
 import 'package:framy_generator/framy_object.dart';
 import 'package:framy_generator/generator/app_bar_generator.dart';
 import 'package:framy_generator/generator/app_bar_page_generator.dart';
+import 'package:framy_generator/generator/available_constructor_names_generator.dart';
 import 'package:framy_generator/generator/button_page_generator.dart';
 import 'package:framy_generator/generator/colors_page_generator.dart';
+import 'package:framy_generator/generator/constructor_dropdown_generator.dart';
 import 'package:framy_generator/generator/datetime_dependency_input_generator.dart';
 import 'package:framy_generator/generator/dependency_model_generator.dart';
 import 'package:framy_generator/generator/drawer_generator.dart';
@@ -82,11 +84,13 @@ class FramyGenerator extends GeneratorForAnnotation<FramyApp> {
     buffer.writeln(generateDateTimeDependencyInput());
     buffer.writeln(generateWidgetListDependencyInput(modelFramyObjects));
     buffer.writeln(generatePresetDropdown());
+    buffer.writeln(generateConstructorDropdown());
     buffer.writeln(
         '\n// ======================== MAPS etc ===========================\n');
     buffer.writeln(generateModelConstructorMap(modelFramyObjects));
     buffer.writeln(generateEnumMap(modelFramyObjects));
     buffer.writeln(generateSubDependenciesMap(modelFramyObjects));
+    buffer.writeln(generateAvailableConstructorNames(modelFramyObjects));
     buffer.writeln(generatePresets(presetFramyObjects));
     return buffer.toString();
   }