+++ framy_generator/test/generators/widget_dependencies_panel_generator_test.dart
@@ -24,8 +24,7 @@ void main() {
     test('should contain onChanged callback', () {
       final result = generateWidgetDependenciesPanel();
       expect(
-        result.contains(
-            'final void Function(String name, dynamic value) onChanged;'),
+        result.contains('final ValueChanged<FramyDependencyModel> onChanged;'),
         isTrue,
       );
     });