+++ test_apps/counter_app/lib/main.app.framy.dart
@@ -795,8 +795,8 @@ class _FramyCustomPageState extends State<FramyCustomPage> {
 
   dynamic dependencyValue(String name) => dependency(name).value;
 
-  void onChanged(String name, dynamic dependencyValue) {
-    setState(() => dependency(name).value = dependencyValue);
+  void onChanged(FramyDependencyModel dependencyModel) {
+    setState(() {});
   }
 
   @override
@@ -895,16 +895,18 @@ class FramyDependencyModel<T> {
   final String name;
   final String type;
   T value;
-  final List<FramyDependencyModel> subDependencies;
   T lastCustomValue;
+  String constructor;
+  List<FramyDependencyModel> subDependencies;
 
-  FramyDependencyModel(this.name, this.type, this.value, this.subDependencies)
+  FramyDependencyModel(this.name, this.type, this.value, this.subDependencies,
+      {this.constructor = ''})
       : lastCustomValue = value;
 }
 
 class FramyWidgetDependenciesPanel extends StatelessWidget {
   final List<FramyDependencyModel> dependencies;
-  final void Function(String name, dynamic value) onChanged;
+  final ValueChanged<FramyDependencyModel> onChanged;
   final Map<String, Map<String, dynamic>> presets;
 
   const FramyWidgetDependenciesPanel(
@@ -939,7 +941,7 @@ class FramyWidgetDependenciesPanel extends StatelessWidget {
 
 class FramyWidgetDependenciesFAB extends StatelessWidget {
   final List<FramyDependencyModel> dependencies;
-  final void Function(String name, dynamic value) onChanged;
+  final ValueChanged<FramyDependencyModel> onChanged;
   final Map<String, Map<String, dynamic>> presets;
 
   const FramyWidgetDependenciesFAB(
@@ -966,9 +968,9 @@ class FramyWidgetDependenciesFAB extends StatelessWidget {
                 child: FramyWidgetDependenciesPanel(
                   dependencies: dependencies,
                   presets: presets,
-                  onChanged: (s, v) {
+                  onChanged: (dep) {
                     setState(() {});
-                    onChanged(s, v);
+                    onChanged(dep);
                   },
                 ),
               ),
@@ -993,7 +995,7 @@ InputDecoration get _framyInputDecoration => InputDecoration(
 
 class FramyWidgetDependencyInput extends StatelessWidget {
   final FramyDependencyModel dependency;
-  final void Function(String name, dynamic value) onChanged;
+  final ValueChanged<FramyDependencyModel> onChanged;
   final Map<String, Map<String, dynamic>> presets;
   final Widget trailing;
 
@@ -1005,7 +1007,8 @@ class FramyWidgetDependencyInput extends StatelessWidget {
     if (value != null && !isValueAPreset(presets, dependency.type, value)) {
       dependency.lastCustomValue = value;
     }
-    onChanged(dependency.name, value);
+    dependency.value = value;
+    onChanged(dependency);
   }
 
   @override
@@ -1037,6 +1040,13 @@ class FramyWidgetDependencyInput extends StatelessWidget {
             ],
           ),
         ),
+        if (!isDependencyAPreset(presets, dependency) &&
+            framyAvailableConstructorNames.containsKey(dependency.type) &&
+            framyAvailableConstructorNames[dependency.type].length > 1)
+          FramyConstructorDropdown(
+            dependency: dependency,
+            onChanged: onChanged,
+          ),
         if (!isDependencyAPreset(presets, dependency))
           if (dependency.type == 'bool')
             InputDecorator(
@@ -1162,8 +1172,7 @@ class FramyModelInput extends StatelessWidget {
                       child: FramyWidgetDependencyInput(
                         dependency: dep,
                         presets: presets,
-                        onChanged: (name, value) {
-                          dependency(name).value = value;
+                        onChanged: (changedDep) {
                           onChanged(dependencies);
                         },
                       ),
@@ -1178,7 +1187,7 @@ class FramyModelInput extends StatelessWidget {
 
 class FramyDateTimeDependencyInput extends StatelessWidget {
   final FramyDependencyModel dependency;
-  final void Function(dynamic value) onChanged;
+  final ValueChanged<dynamic> onChanged;
   final Map<String, Map<String, dynamic>> presets;
 
   const FramyDateTimeDependencyInput(
@@ -1242,9 +1251,14 @@ class FramyWidgetListDependencyInput extends StatelessWidget {
                 listType,
                 dependency.value[i],
                 dependency.subDependencies[i].subDependencies,
+                constructor: dependency.subDependencies[i].constructor,
               ),
-              onChanged: (name, val) {
-                dependency.value[i] = val;
+              onChanged: (changedDep) {
+                dependency.value[i] = changedDep.value;
+                dependency.subDependencies[i].constructor =
+                    changedDep.constructor;
+                dependency.subDependencies[i].subDependencies =
+                    changedDep.subDependencies;
                 onChanged(dependency);
               },
               presets: presets,
@@ -1344,6 +1358,50 @@ class FramyPresetDropdown extends StatelessWidget {
   }
 }
 
+class FramyConstructorDropdown extends StatelessWidget {
+  final FramyDependencyModel dependency;
+  final ValueChanged<FramyDependencyModel> onChanged;
+
+  FramyConstructorDropdown({this.dependency, this.onChanged})
+      : super(
+          key: Key('framy_dependency_${dependency.name}_constructor_dropdown'),
+        );
+
+  @override
+  Widget build(BuildContext context) {
+    return InputDecorator(
+      decoration: InputDecoration(
+        labelText: 'Constructor',
+        isDense: true,
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(8),
+          borderSide: BorderSide(),
+        ),
+      ),
+      child: DropdownButtonHideUnderline(
+        child: DropdownButton<String>(
+          isDense: true,
+          value: dependency.constructor,
+          onChanged: (conName) {
+            dependency.constructor = conName;
+            dependency.subDependencies =
+                createSubDependencies(dependency.type, dependency.constructor);
+            onChanged(dependency);
+          },
+          items: framyAvailableConstructorNames[dependency.type]
+              .map((constructorName) => DropdownMenuItem<String>(
+                    value: constructorName,
+                    child: Text(constructorName.isEmpty
+                        ? 'Default'
+                        : constructorName.replaceAll('.', '')),
+                  ))
+              .toList(),
+        ),
+      ),
+    );
+  }
+}
+
 // ======================== MAPS etc ===========================
 
 final framyModelConstructorMap =
@@ -1359,11 +1417,14 @@ final framyModelConstructorMap =
 final framyEnumMap = <String, List<dynamic>>{
   'MaterialTapTargetSize': MaterialTapTargetSize.values,
 };
-List<FramyDependencyModel> createSubDependencies(String type) {
-  switch (type) {
+List<FramyDependencyModel> createSubDependencies(String type,
+    [String constructor = '']) {
+  switch (type + constructor) {
     default:
       return [];
   }
 }
+
+Map<String, List<String>> framyAvailableConstructorNames = {};
 
 Map<String, Map<String, dynamic>> createFramyPresets() => {};