+++ test_apps/counter_app/test/widget_dependencies_panel_test.dart
@@ -42,8 +42,8 @@ void main() {
       await tester.pumpWidget(TestMaterialAppWithScaffold(
         FramyWidgetDependenciesPanel(
           dependencies: [model],
-          onChanged: (name, val) =>
-              model = FramyDependencyModel<String>(name, 'String', val, []),
+          onChanged: (dep) => model =
+              FramyDependencyModel<String>(dep.name, 'String', dep.value, []),
           presets: {},
         ),
       ));