+++ framy_generator/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## [0.1.3]
 
+* Support for models with multiple constructors
 * Added usage of DevicePreview 
 
 ## [0.1.2]